requirements.txt
blinker==1.9.0
branca==0.8.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.1.0
folium==0.18.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
k==0.0.1
MarkupSafe==3.0.2
numpy==2.1.3
packaging==24.2
pillow==11.0.0
pytesseract==0.3.13
requests==2.32.3
urllib3==2.2.3
Werkzeug==3.1.3
xyzservices==2024.9.0
gunicorn
